+++ components/compress/compressor_core.tsx
@@ -6,7 +6,6 @@ import InvalidImageType from "../error/InvalidImgType";
 import Button from "../buttons&links/Button";
 import Image from "next/image";
 import { FaLock, FaLockOpen } from "react-icons/fa";
-import { dataConverter, isSupportedFileType, round } from "@/utils/utils";
 
 interface CompressorCoreProps {
   file: File;
@@ -34,6 +33,29 @@ function CompressorCore({ file, removeSelf, inputImageSrc, setOutputImageSrc}: C
   const [aspectRatioLocked, setAspectRatioLocked] = useState<boolean>(false);
   const [aspectRatioW, setAspectRatioW] = useState<number>(1);
   const [aspectRatioH, setAspectRatioH] = useState<number>(1);
+
+  const dataConverter = (num: number, divide: number): string => {
+    if (num / divide > 1024) {
+      const toMb = `${num / divide / 1024}`;
+      return `${parseInt(toMb)}MB`;
+    }
+    if (num / divide < 1024) {
+      const toKb = `${num / divide}`;
+      return `${parseInt(toKb)}KB`;
+    }
+    return "";
+  };
+
+  const round = (num: number, r: number = 0): number => {
+    const pow = Math.pow(10, r);
+    return Math.floor(num * pow) / pow;
+  };
+
+  const isSupportedFileType = (file: File | null): boolean => {
+    if (!file) return false;
+    const supportedFileTypes = ["image/jpeg", "image/jpg", "image/png"];
+    return supportedFileTypes.includes(file.type);
+  }
 
   useEffect(() => {
     if (aspectRatioLocked || !W || !H) return;
